refactor(core): extract input bindings and canvas export
